Move loading spinner to inside Refresh
